extract readFolder to file
